Restructuring Presenters
Move presenters into appropriate locations
