+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/Presenter.java
@@ -1,0 +1,9 @@
+package edu.byu.cs.tweeter.client.presenter;
+
+public class Presenter {
+  View v;
+
+  public Presenter(View v) {
+    this.v = v;
+  }
+}
