+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/View.java
@@ -1,0 +1,5 @@
+package edu.byu.cs.tweeter.client.presenter;
+
+public interface View {
+
+}
