+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/paged/FeedPresenter.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import edu.byu.cs.tweeter.client.model.service.FeedService;
 import edu.byu.cs.tweeter.client.model.service.UserService;
-import edu.byu.cs.tweeter.client.presenter.PagedView;
 import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
