+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/paged/PagedView.java
@@ -1,11 +1,12 @@
-package edu.byu.cs.tweeter.client.presenter;
+package edu.byu.cs.tweeter.client.presenter.paged;
 
 import java.net.MalformedURLException;
 import java.util.List;
 
+import edu.byu.cs.tweeter.client.presenter.View;
 import edu.byu.cs.tweeter.model.domain.User;
 
-public interface PagedView<T>{
+public interface PagedView<T> extends View {
   void navigateToUser(User user);
 
   void setLoading(boolean value) throws MalformedURLException;
